Remove context class
