--- src/Microsoft.AspNet.DependencyInjection/NestedProviders/NestedProviderContext.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.AspNet.DependencyInjection
-{
-    public class NestedProviderContext<T>
-    {
-        public T Result { get; set; }
-    }
-}
